ArtRay.Geometry

Documentation

quadSolve :: Double -> Double -> Double -> (Double, Double)

Solve a quadratic. Returns NaN for non-real roots, and two equal values for double roots.

firstIntersection :: Ray -> Primitive -> Maybe (Double, Vec3, Primitive)

Find the first intersection between a primitive and a ray. Returns Nothing if the ray never intersects the primitive.

pointToRay' :: Viewer -> Point2D -> Ray

Find the ray for a point on the image plane

farpoint :: Ray -> Vec3 -> Double -> Vec3

Finds the point on a ray whose projection on a vector has a specific length.

deformRay :: Scene -> Ray -> (Double, Double) -> Ray

Deform a ray by (i, j) in the image plane, maintaining the focal point of the ray.

pointToRay :: Scene -> Viewer -> Point2D -> [Ray]

Find a family of rays that converge at the focal distance for a point in the image plane.

reflectAbout :: Vec3 -> Vec3 -> Vec3

Reflect the first vector about the second

refractVector :: Double -> Vec3 -> Vec3 -> Vec3

Refract the second vector around the first with the provided refraction index.

normal :: Primitive -> Vec3 -> Vec3

Find the normal vector at a point on a primitive. Note that this function does not require that the point be actually incident to the primitive.