ArtRay.Primitives

Documentation

data Scene

Constructors

Scene

Fields

background :: ColorTriple

options :: [Option]

geom :: [Primitive]

lights :: [Light]

viewer :: Viewer

data Option

Constructors

Antialiased Double

GlobalAmbient ColorTriple

data Primitive

Constructors

Sphere

Fields

center :: Vec3

radius :: Double

material :: Material

Plane

Fields

pnorm :: Vec3

point :: Vec3

material :: Material

RectPrism

Fields

corner :: Vec3

x :: Vec3

y :: Vec3

z :: Vec3

material :: Material

data Ray

Constructors

Ray

Fields

direction :: Vec3

position :: Vec3

type ColorTriple = (Double, Double, Double)

data Light

Constructors

PhongLight

Fields

speclight :: ColorTriple

difflight :: ColorTriple

loclight :: Vec3

data CombinationModel

Constructors

WeightSum

Fields

weight :: Double

FlatSum

Multiply

data Material

Constructors

ColorMaterial

Fields

basecolor :: ColorTriple

ReflectiveMaterial

Fields

base :: Material

reflectivity :: Double

PhongMaterial

Fields

specular :: ColorTriple

diffuse :: ColorTriple

ambient :: ColorTriple

phongexp :: Int

TransparentMaterial

Fields

base :: Material

cmodel :: CombinationModel

refindex :: Double

NullMaterial

data Viewer

Constructors

Viewer

Fields

location :: Vec3

The location of the viewer

u :: Vec3

A vector pointing horizontally along the image plane

v :: Vec3

A vector pointing vertically along the image plane

f :: Vec3

A vector pointing from the viewer to the center of the image plane

data Point2D

Constructors

Point2D Double Double
Describes a point in the image using pixel coordinates, allowing for | fractional pixels for subpixel sampling.

RelPoint2D Double Double
Describes a point in the image using offsets from the center, where each field goes from zero to one.

transmittance :: Material -> Double

glambient :: Scene -> ColorTriple

subpixels :: Scene -> Maybe Double

p2d :: Int -> Int -> Point2D

color :: Double -> Double -> Double -> ColorTriple

colorm :: Double -> Double -> Double -> Material

toRelPoint :: Size -> Point2D -> Point2D

view

Arguments

:: Vec3
The location of the viewer

-> Double
The field of view of the viewer, in radians

-> Vec3
The vector from viewer to center of image plane

-> Vec3
Unit vector pointing vertically along the image plane

-> Viewer

Creates a viewer

colorFrom :: ColorTriple -> Color

combine :: ColorTriple -> ColorTriple -> ColorTriple

scale :: ColorTriple -> Double -> ColorTriple

weightedCombine :: Double -> ColorTriple -> ColorTriple -> ColorTriple

sumLight :: [ColorTriple] -> ColorTriple

sumColor :: ColorTriple -> ColorTriple -> ColorTriple

meanColor :: [ColorTriple] -> ColorTriple