ArtRay.RayTrace

Documentation

pixelColor :: Size -> Scene -> Viewer -> Point2D -> ColorTriple

Get the color for a pixel in the scene

colorFor

Arguments

:: Scene
The scene we're operating within

-> Primitive
The shape to determine the color for

-> Material
The material to examine (not necessarily the result of (material shape)

-> Vec3
The incident vector of the ray

-> Vec3
The location of intersection

-> Int
The stack depth

-> ColorTriple
The color at that point

Find the color at a pixel defined by the point

Finds the color at a point on a primitive, using the provided material. Both the primitive and the material must be supplied, as some primitives may have multiple materials that need to be calculated. The stack depth of ray tracing is tracked but currently is unused and unlimited.

occluded :: Scene -> Primitive -> Vec3 -> Vec3 -> Double

Returns how much occlusion occurs between two points x1 and x2. The shape at x1 must be provided to ensure that it is excluded when calculating intersections. The return value ranges from 0 to 1, where 0 is fully occluded and 1 is not occluded at all. Intermediate values are possible because of semi-transparent material that may be between x1 and x2.

phongLight

Arguments

:: Scene

-> Primitive

-> Material

-> Ray
The incoming ray, whose location is the point of intersection

-> Vec3
The surface normal at the illumination point

-> Light

-> ColorTriple

Calculates the illumination of a point on a primitive with a Phong material.

colorAtRay :: Scene -> Int -> Ray -> ColorTriple

Find the color of the first object that intersects a ray

colorAtRay' :: Scene -> Int -> Ray -> [Primitive] -> ColorTriple

Find the color of the first object that intersects a ray, excluding the provided primitives from the intersection.

orderTuples :: (Double, Vec3, Primitive) -> (Double, Vec3, Primitive) -> Ordering

Order tuples representing intersections, from first to last intersection

geomAtRay :: Scene -> Ray -> [Primitive] -> Maybe (Vec3, Primitive)

Find the first primitive at a ray

intersectWithScene :: Scene -> Ray -> [Primitive] -> [(Double, Vec3, Primitive)]

Find all points of intersections of a ray with the scene