TrayRacer.Primitives

Documentation

data Primitive

Constructors

Sphere Vec3 Double Color

data Ray

Constructors

Ray

Fields

direction :: Vec3

position :: Vec3

data Point2D

Constructors

Point2D Int Int
Describes a point in the image using pixel coordinates

RelPoint2D Double Double
Describes a point in the image using offsets from the center, where each field goes from zero to one.

data Viewer

Constructors

Viewer

Fields

location :: Vec3

The location of the viewer

u :: Vec3

A vector pointing horizontally along the image plane

v :: Vec3

A vector pointing vertically along the image plane

f :: Vec3

A vector pointing from the viewer to the center of the image plane

view

Arguments

:: Vec3
The location of the viewer

-> Double
The field of view of the viewer, in radians

-> Vec3
The vector from viewer to center of image plane

-> Vec3
Unit vector pointing vertically along the image plane

-> Viewer

Creates a viewer

data Scene

Constructors

Scene [Primitive]