ArtRay.Primitives

Documentation

data Scene

The root object of any rendered scene in ArtRay

Constructors

Scene

Fields

background :: ColorTriple

The background color of the scene -- displayed whenever a ray doesn't intersect any geometry.

options :: [Option]

A list of rendering options

geom :: [Primitive]

The geometry of the scene

lights :: [Light]

The lights in the scene

viewer :: Viewer

Camera information

data Option

Rendering-related options for scenes

Constructors

GlobalAmbient ColorTriple
The global ambient light. If provided, this will illuminate all primitives with nonzero ambient material response at all points.

Antialiased
Antialiasing options. When the Antialiased option is provided with msaaSamples > 1, ArtRay will use MSAA subpixel sampling to antialias the image. Note that multiplying msaaSamples by n increases runtime by O(n^2).

Fields

msaaSamples :: Double

DepthOfField
Depth of field options. If this option is not provided, the camera has infinite depth of field. When this option is provided, the camera simulates a real camera with focal length and aperture. Aperture is a number, usually on the order of 1, which controls how tight the depth of field is -- higher aperture values lead to narrower areas in focus. The focal length option determines how far away from the camera's location is in focus. The samples parameter is the square root of the number of rays shot out to determine the color at each pixel. For simple scenes, this can be as low as 2. The run time scales as O(samples^2).

Fields

focalLength :: Double

aperture :: Double

dofSamples :: Int

data Primitive

Constructors

Sphere

Fields

center :: Vec3

radius :: Double

material :: Material

Plane

Fields

pnorm :: Vec3

point :: Vec3

material :: Material

data Ray

Constructors

Ray

Fields

direction :: Vec3

position :: Vec3

type ColorTriple = (Double, Double, Double)

data Light

Constructors

PhongLight

Fields

speclight :: ColorTriple

difflight :: ColorTriple

loclight :: Vec3

data CombinationModel

Combination models determine how materials on an object are combined.

Constructors

WeightSum
Sum the two colors after multiplying the primary by weight and the secondary by (1-weight). The weighting term should be in [0,1].

Fields

weight :: Double

FlatSum
Sum the colors, truncating if any component goes over the maximum.

Multiply
Take the product of the two colors. This creates weird effects which you almost certainly do not want, but it's interesting to look at.

data Material

Constructors

ColorMaterial
A material where every point on the material has the same color.

Fields

basecolor :: ColorTriple

ReflectiveMaterial
A material which reflects rays. Optionally, the material can have a base material, whose color is blended with the reflections using the reflectivity parameter. TODO: Use a combination model here.

Fields

base :: Material

reflectivity :: Double

PhongMaterial
A Phong-reflectance based material.

Fields

specular :: ColorTriple

diffuse :: ColorTriple

ambient :: ColorTriple

phongexp :: Int

TransparentMaterial
A transparent material with a refractive index. This can be blended with a base material using the provided combination model.

Fields

base :: Material

cmodel :: CombinationModel

refindex :: Double

NullMaterial
A material that doesn't exist. This can be convenient for a placeholder when you don't want to provide an actual material (like as the base material of a TransparentMaterial that is fully transparent).

data Viewer

Constructors

Viewer

Fields

location :: Vec3

The location of the viewer

u :: Vec3

A vector pointing horizontally along the image plane

v :: Vec3

A vector pointing vertically along the image plane

f :: Vec3

A vector pointing from the viewer to the center of the image plane

data Point2D

Constructors

Point2D Double Double
Describes a point in the image using pixel coordinates, allowing for fractional pixels for subpixel sampling.

RelPoint2D Double Double
Describes a point in the image using offsets from the center, where each field goes from zero to one.

transmittance :: Material -> Double

Returns the transmittance (1 minus the opacity) of any material.

glambient :: Scene -> ColorTriple

Returns the global ambient of the scene

subpixels :: Scene -> Maybe Double

Returns the number of subpixels used for MSAA, or Nothing if no antialiasing is to be performed.

dofinfo :: Scene -> Maybe Option

Returns the DepthOfField option, or nothing if no depth of field is to be used.

dofenabled :: Scene -> Bool

Returns True if depth of field is to be simulated, and false otherwise.

dofsamples :: Scene -> Int

Returns the number of samples to use for calculating depth of field. This will throw an exception if depth of field is not enabled for the provided scene.

dofdepth :: Scene -> Double

Returns the focal length of the camera. This will throw an exception if depth of field is not enabled for the provided scene.

dofaperture :: Scene -> Double

Returns the aperture size of the camera. This will throw an exception if depth of field is not enabled for the provided scene.

p2d :: Int -> Int -> Point2D

Convenience method for turning a pair of integers into a Point2D.

color :: Double -> Double -> Double -> ColorTriple

Convenience method for creating colors

colorm :: Double -> Double -> Double -> Material

Convenience method for creating ColorMaterials.

toRelPoint :: Size -> Point2D -> Point2D

Converts pixel locations into relative locations in [-1,1], where -1 corresponds to the left or bottom edge, and 1 corresponds to the right or top edge.

view

Arguments

:: Vec3
The location of the viewer

-> Double
The field of view of the viewer, in radians

-> Vec3
The vector from viewer to center of image plane

-> Vec3
Unit vector pointing vertically along the image plane

-> Viewer

Creates a viewer

colorFrom :: ColorTriple -> Color

Convert an ArtRay color to a GD color.

combine :: ColorTriple -> ColorTriple -> ColorTriple

Take the product of two colors

scale :: ColorTriple -> Double -> ColorTriple

Multiply a color by a scalar

weightedCombine :: Double -> ColorTriple -> ColorTriple -> ColorTriple

Weight two colors with the provided weighting factor. The first color is multiplied by the weight, the second by one minus the weight.

sumLight :: [ColorTriple] -> ColorTriple

Ensure that each component of the color is in [0, 1].

Take the sum of a list of colors

sumColor :: ColorTriple -> ColorTriple -> ColorTriple

Take the sum of a pair of colors, truncating if it results in a value outside [0, 1].

sumColor' :: ColorTriple -> ColorTriple -> ColorTriple

Take the sum of a pair of colors without normalizing.

meanColor :: [ColorTriple] -> ColorTriple

Find the mean value of a list of colors.